Method to turn a path into a PV name
In the AA storage, a PV name is broken up into parts
and there are subdirectories with each part of the name.
This method is to reverse that process so you can work out
what a PV name is from the path to the PB file.

The first three slashes in the path become dashes in the
PV anme. Subsequent slashes become colons.

E.g. "BL14I/MO/STAGE/01/XF.RBV:2019.pb" goes to
"BL14I-MO-STAGE-01:XF.RBV"