Fix bugs in display of errors
Replaced the expression
     [[]] * rows
to initialize the errors list because it has unexpected effects;
in this case the list in each element is the same object
so appending to one appends to all the others!

In this way each element in the outer list was getting an item
appended for every event.

Also fix bug to only show "No events" when there are 0 events, not
1.