Make ordering of URL params repeatable
Addition of `pv` argument to `RestClient.get_all_pvs` showed up
a problem in `_construct_url` that ordering of URL parameters
was not repeatable because they are got from the kwargs as a dict.

We get around this by sorting by key and storing as a list of
tuples of (key, value). We assume that no-one cares what the
order of URL parameters is, as long as it is deterministic.